% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{aggregated}
\alias{aggregated}
\title{Aggregated Data Set}
\format{A data frame with 350 observations on the following 9 variables:
\describe{
\item{state}{The name of the state}
\item{year}{The year, which spans from 2011 to 2017}
\item{gdp}{GDP per capita for the state, given in chained 2009 dollars}
\item{gini}{The Gini index for the state, which ranges from 0 (perfect equality) to 1 (perfect inequality)}
\item{number_homeless}{The homeless population for the state}
\item{population}{The total population of the state}
\item{rentincome}{The median gross rent as a share of household income for the state}
\item{unemployment}{The average unemployment rate for the state}
\item{share_homeless}{The homeless population as a share of the total population}
}}
\usage{
aggregated
}
\description{
A concatenated version of all data sets available in the `homelessness` package.
}
\examples{
summary(aggregated)
}
\keyword{datasets}
