% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gdp}
\alias{gdp}
\title{GDP Per Capita Data Set}
\format{A data frame with 408 observations on the following 3 variables:
\describe{
\item{state}{The name of the state}
\item{year}{The year, which spans from 2011 to 2018}
\item{gdp}{GDP per capita for the state, given in chained 2009 dollars}
}}
\source{
\url{https://en.wikipedia.org/wiki/List_of_U.S._states_by_GDP_per_capita}
}
\usage{
gdp
}
\description{
The data includes GDP per capita estimates for all U.S. states plus the District of Columbia. The data was sourced from Wikipedia.
}
\examples{
hist(gdp$gdp)

}
\keyword{datasets}
