% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{homeless}
\alias{homeless}
\title{Homelessness Data Set}
\format{A data frame with 600 observations on the following 3 variables:
\describe{
\item{state}{The name of the state}
\item{year}{The year, which spans from 2007 to 2018}
\item{number_homeless}{The homeless population for the state}
}}
\source{
\url{https://www.hudexchange.info/resource/5783/2018-ahar-part-1-pit-estimates-of-homelessness-in-the-us/}
}
\usage{
homeless
}
\description{
The data includes the number of homeless persons for each U.S. state. The data was sourced from the United States Department of Housing and Urban Development.
}
\examples{
hist(homeless$number_homeless)
}
\keyword{datasets}
