% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{population}
\alias{population}
\title{Population Data Set}
\format{A data frame with 612 observations on the following 3 variables:
\describe{
\item{state}{The name of the state}
\item{year}{The year, which spans from 2007 to 2018}
\item{population}{The total population of the state}
}}
\source{
\url{https://www.census.gov/programs-surveys/acs}
}
\usage{
population
}
\description{
The data includes the total population for each U.S. state plus the District of Columbia. The data was sourced from the U.S. Census Bureau's American Community Survey.
}
\examples{
mean(population$population)
}
\keyword{datasets}
